Move teams service to new directory
